refactor: product 모듈 분리
