+++ build.gradle
@@ -44,9 +44,20 @@ allprojects {
 
 dependencies {
     implementation project(':kitchenpos-common')
+    implementation project(':kitchenpos-product')
+
+    testCompile project(':kitchenpos-common').sourceSets.test.output
+    testCompile project(':kitchenpos-product').sourceSets.test.output
 }
 
 project(':kitchenpos-common') {
     dependencies {
+    }
+}
+
+project(':kitchenpos-product') {
+    dependencies {
+        implementation project(':kitchenpos-common')
+        testCompile project(':kitchenpos-common').sourceSets.test.output
     }
 }
