+++ kitchenpos-common/src/test/java/kitchenpos/common/domain/NameTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.common.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
