+++ kitchenpos-product/src/main/resources/db/migration/V3__alter_project_tables.sql
@@ -1,0 +1,2 @@
+ALTER TABLE order_line_item ADD COLUMN menu_name VARCHAR(255) NOT NULL;
+ALTER TABLE order_line_item ADD COLUMN menu_price DECIMAL(19, 2) NOT NULL;
