+++ kitchenpos-product/src/main/resources/db/sql/truncate.sql
@@ -1,0 +1,10 @@
+SET FOREIGN_KEY_CHECKS = 0;
+TRUNCATE TABLE MENU;
+TRUNCATE TABLE MENU_GROUP;
+TRUNCATE TABLE MENU_PRODUCT;
+TRUNCATE TABLE ORDERS;
+TRUNCATE TABLE ORDER_LINE_ITEM;
+TRUNCATE TABLE ORDER_TABLE;
+TRUNCATE TABLE PRODUCT;
+TRUNCATE TABLE TABLE_GROUP;
+SET FOREIGN_KEY_CHECKS = 1;
