+++ settings.gradle
@@ -1,2 +1,4 @@
 rootProject.name = 'kitchenpos'
 include 'kitchenpos-common'
+include 'kitchenpos-product'
+
