+++ src/test/java/kitchenpos/order/acceptanceTest/OrderAcceptanceTest.java
@@ -1,12 +1,12 @@
 package kitchenpos.order.acceptanceTest;
 
-import static kitchenpos.menugroup.acceptance.MenuGroupRestAssured.메뉴_그룹_생성_요청;
 import static kitchenpos.menu.acceptance.MenuRestAssured.메뉴_생성_요청;
+import static kitchenpos.menugroup.acceptance.MenuGroupRestAssured.메뉴_그룹_생성_요청;
 import static kitchenpos.order.acceptanceTest.OrderRestAssured.주문_목록_조회_요청;
 import static kitchenpos.order.acceptanceTest.OrderRestAssured.주문_상태_수정_요청;
 import static kitchenpos.order.acceptanceTest.OrderRestAssured.주문_생성_요청;
-import static kitchenpos.product.acceptance.ProductRestAssured.상품_생성_요청;
 import static kitchenpos.ordertable.acceptance.TableRestAssured.주문_테이블_생성_요청;
+import static kitchenpos.product.acceptance.ProductRestAssured.상품_생성_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
@@ -17,13 +17,13 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.AcceptanceTest;
-import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.common.acceptance.AcceptanceTest;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.menugroup.dto.MenuGroupRequest;
 import kitchenpos.menugroup.dto.MenuGroupResponse;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
