+++ src/test/java/kitchenpos/tablegroup/application/TableGroupServiceTest.java
@@ -13,7 +13,6 @@ import java.util.List;
 import java.util.Optional;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuProducts;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItems;
